flat evasion/accuracy stats/spell pen fixes
Fixed syntax for spell penetration, unsure if im supposed to be using target. or not, fixed adding flat evade/crit to attack formulas, going to add more for accuracy and stuff later, so when we put static values on gear, its already in there. Also how do i add these to classes?

There is one "unexpected end" error on the statuses thingy, but the game runs, idk how to fix that.